+++ src/app/features/pokemon/application/services/get-pokemon-list/get-pokemon-list.ts
@@ -1,6 +1,6 @@
 import { RequestResponse } from '@/core/application/http-response/http-response'
-import type { HttpClient } from '@/core/application/protocols'
-import { HttpMethod } from '@/core/application/protocols'
+import type { HttpClient } from '@/core/application/protocols/http-client'
+import { HttpMethod } from '@/core/application/protocols/http-client'
 import { type ServiceCommand } from '@/core/domain/command/service-command'
 import { error, success } from '@/core/domain/either/either'
 import type { PokemonListItem } from '@/app/features/pokemon/domain/models'
