--- src/core/application/protocols/index.ts
@@ -1,2 +1,0 @@
-export * from './cache-storage'
-export * from './http-client'
