--- src/core/domain/exceptions/index.ts
@@ -1,3 +1,0 @@
-export { UnexpectedError } from './unexpected-error'
-export { AccessDeniedError } from './access-denied-error'
-export type { DomainException } from './domain-exception'
