--- src/core/infra/index.ts
@@ -1,2 +1,0 @@
-export * from './cache'
-export * from './http'
