+++ src/core/query/error-handler.ts
@@ -1,4 +1,4 @@
-import type { DomainException } from '@/core/domain/exceptions'
+import type { DomainException } from '@/core/domain/exceptions/domain-exception'
 
 export function globalQueryErrorHandler(err: unknown) {
   const e = err as Partial<DomainException>
