--- src/core/query/index.ts
@@ -1,4 +1,0 @@
-export * from './client'
-export * from './service-query'
-export * from './service-mutation'
-export * from './service-infinite-query'
